Added sample permission for positions module

+++ app/Http/Controllers/User/PositionController.php
@@ -29,6 +29,8 @@ class PositionController extends Controller
      */
     public function index(): JsonResponse
     {
+        $this->authorize( 'position_show' );
+
         $positions = PositionModel::all()->toArray();
 
         return $this->result->success(array_reverse($positions));
@@ -39,6 +41,8 @@ class PositionController extends Controller
      */
     public function get($id): JsonResponse
     {
+        $this->authorize( 'position_show' );
+
         $position = PositionModel::find($id);
 
         return $this->result->success($position);
@@ -49,6 +53,8 @@ class PositionController extends Controller
      */
     public function store(PositionRequest $request): JsonResponse
     {
+        $this->authorize( 'position_create' );
+
         $position = new PositionModel(['name' => $request->input('name')]);
         $position->save();
 
@@ -60,6 +66,7 @@ class PositionController extends Controller
      */
     public function update(PositionRequest $request, $id): JsonResponse
     {
+        $this->authorize( 'position_edit' );
 
         $position = PositionModel::find($id);
         $position->update($request->all());
@@ -72,6 +79,8 @@ class PositionController extends Controller
      */
     public function destroy($id): JsonResponse
     {
+        $this->authorize( 'position_delete' );
+
         $position = PositionModel::find($id);
         $position->delete();
